Create game_logi and player files
