+++ .gitignore
@@ -1,2 +1,1 @@
 /trials.rb
-/lib/game_logic.rb
