+++ bin/main.rb
@@ -1,98 +1,8 @@
 #!/usr/bin/env ruby
-
-# Handles all user interaction messages
+require_relative "../lib/player.rb"
+require_relative "../lib/game_logic.rb"
 def print_user_message(msg)
   print msg
-end
-
-class TicTacToeGame
-  def initialize
-    @board = Board.new
-    @player_a = Player.new(board: @board, name: "Player A", piece: :X)
-    @player_b = Player.new(board: @board, name: "Player B", piece: :O)
-    @current_player = @player_a
-  end
-
-  def get_player_profile
-    @player_a.get_player_name(@player_a.name)
-    @player_a.get_player_symbol(@player_a.name)
-    @player_b.get_player_name(@player_b.name)
-    @player_b.get_player_symbol(@player_b.name)
-  end
-
-  def play_game
-    @board.render_board
-    get_player_profile
-    loop do
-      @current_player.obtain_coordinates
-      @board.render_board
-      break if game_over?
-
-      switch_players
-    end
-  end
-
-  def game_over?
-    if @board.winning?(@current_player.piece)
-      puts "Congratulations #{@current_player.name}, you have won!"
-    elsif @board.filled?
-      puts "Oop! You have drawn!"
-    else return false
-    end
-    true
-  end
-
-  def switch_players
-    @current_player = @current_player == @player_a ? @player_b : @player_a
-  end
-end
-
-class Player
-  attr_reader :name, :piece
-
-  def initialize(piece:, board:, name: "Anonymous")
-    @name = name
-    @piece = piece
-    @board = board
-  end
-
-  def get_player_name(player_n)
-    loop do
-      print "#{player_n} enter your name as a text without a space or symbol: "
-      name = gets.chomp
-      if !(name =~ /^[0-9a-zA-Z]+$/).nil?
-        @name = name.capitalize # capitalize the first letter of the name
-        break
-      else
-        puts "Invalid name format, try again!"
-      end
-    end
-  end
-
-  def get_player_symbol(player_n)
-    loop do
-      print "#{player_n} enter your symbol as just one letter: "
-      piece = gets.chomp
-      if !(piece =~ /^[a-zA-Z]$/).nil?
-        @piece = piece.upcase
-        break
-      else
-        puts "Invalid symbol format, try again!"
-      end
-    end
-  end
-
-  def obtain_coordinates
-    puts "#{@name} (#{@piece}), Enter coordinates in form of x,y"
-    loop do
-      coordinates = gets.chomp.split(',').map(&:to_i)
-      if coordinates.size != 2
-        print_user_message("Wrong coordinate format!")
-      else
-        break if @board.add_new_piece(coordinates, @piece)
-      end
-    end
-  end
 end
 
 class Board
